1、新增跳转私有拦截器。 2、新增非Activity跳转支持。 3、新增跳过公有拦截器配置。

+++ app/src/main/java/com/lzp/deeplinkso/demo/TestActivity.kt
@@ -13,7 +13,7 @@ class TestActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_test)
 
-        val userId = intent.getIntExtra("userId", 0)
+        val userId = intent.getLongExtra("userId", 0)
         val userName = intent.getStringExtra("userName")
         Toast.makeText(this, "$userId:$userName", Toast.LENGTH_SHORT).show()
     }
